+++ app/src/main/java/com/boardgamegeek/io/model/GeekListResponse.java
@@ -1,26 +1,14 @@
 package com.boardgamegeek.io.model;
 
-import com.boardgamegeek.model.GeekListComment;
-import com.boardgamegeek.model.GeekListItem;
-
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.ElementList;
 import org.simpleframework.xml.Root;
 
-import java.text.SimpleDateFormat;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Locale;
 
 @Root(name = "geeklist")
 public class GeekListResponse {
-	public static final SimpleDateFormat FORMAT = new SimpleDateFormat("EEE, dd MMM yyyy HH:mm:ss Z", Locale.US);
-
-	@SuppressWarnings("unused")
-	public GeekListResponse() {
-	}
-
 	@Attribute public int id;
 	@Element public String postdate;
 	@Element public String editdate;
@@ -31,14 +19,7 @@ public class GeekListResponse {
 	@Element public String username;
 	@Element public String title;
 	@Element public String description;
-	@ElementList(name = "comment", inline = true, required = false) private ArrayList<GeekListComment> comments;
-	@ElementList(name = "item", inline = true, required = false) private List<GeekListItem> items;
-
-	private ArrayList<GeekListComment> getComments() {
-		return comments;
-	}
-
-	public List<GeekListItem> getItems() {
-		return items;
-	}
+	@ElementList(name = "comment", inline = true, required = false)
+	public List<GeekListComment> comments;
+	@ElementList(name = "item", inline = true, required = false) public List<GeekListItem> items;
 }