+++ app/src/main/java/com/boardgamegeek/ui/adapter/GamePagerAdapter.kt
@@ -6,10 +6,9 @@ import androidx.annotation.DrawableRes
 import androidx.annotation.StringRes
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.FragmentPagerAdapter
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
+import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.boardgamegeek.R
 import com.boardgamegeek.auth.Authenticator
 import com.boardgamegeek.extensions.colorize
@@ -21,8 +20,8 @@ import com.boardgamegeek.ui.dialog.CollectionStatusDialogFragment
 import com.boardgamegeek.ui.viewmodel.GameViewModel
 import com.google.android.material.floatingactionbutton.FloatingActionButton
 
-class GamePagerAdapter(fragmentManager: FragmentManager, private val activity: FragmentActivity, private val gameId: Int, var gameName: String) :
-        FragmentPagerAdapter(fragmentManager, BEHAVIOR_RESUME_ONLY_CURRENT_FRAGMENT) {
+class GamePagerAdapter(private val activity: FragmentActivity, private val gameId: Int, var gameName: String) :
+        FragmentStateAdapter(activity) {
     var currentPosition = 0
         set(value) {
             field = value
@@ -34,6 +33,7 @@ class GamePagerAdapter(fragmentManager: FragmentManager, private val activity: F
     private var heroImageUrl = ""
     private var arePlayersCustomSorted = false
     private var isFavorite = false
+
     @ColorInt
     private var iconColor = Color.TRANSPARENT
     private val tabs = arrayListOf<Tab>()
@@ -50,17 +50,12 @@ class GamePagerAdapter(fragmentManager: FragmentManager, private val activity: F
         updateTabs()
     }
 
-    override fun notifyDataSetChanged() {
-        super.notifyDataSetChanged()
-        updateTabs()
-    }
-
-    override fun getPageTitle(position: Int): CharSequence {
+    fun getPageTitle(position: Int): CharSequence {
         @StringRes val resId = tabs.getOrNull(position)?.titleResId ?: INVALID_RES_ID
         return if (resId != INVALID_RES_ID) activity.getString(resId) else ""
     }
 
-    override fun getItem(position: Int): Fragment {
+    override fun createFragment(position: Int): Fragment {
         return when (tabs.getOrNull(position)?.titleResId) {
             R.string.title_descr -> GameDescriptionFragment.newInstance()
             R.string.title_info -> GameFragment.newInstance()
@@ -73,7 +68,7 @@ class GamePagerAdapter(fragmentManager: FragmentManager, private val activity: F
         }
     }
 
-    override fun getCount(): Int = tabs.size
+    override fun getItemCount(): Int = tabs.size
 
     private fun updateTabs() {
         tabs.clear()
@@ -153,6 +148,7 @@ class GamePagerAdapter(fragmentManager: FragmentManager, private val activity: F
         fab.setTag(R.id.res_id, resId)
     }
 
+    // TODO observe these from the view model
     private fun shouldShowPlays() = Authenticator.isSignedIn(activity) && activity.getSyncPlays()
 
     private fun shouldShowCollection() = Authenticator.isSignedIn(activity) && activity.isCollectionSetToSync()